common: added a new display menu for frac
This work adds a new menu secondary to the display menu.  The new menu
holds the new "frac" display mode.  Fraction input has also been added
by pressing the "dot" key a second time after entering a decimal
number.  In this case, the integer portion will remain an integer, the
fractional portion will become the numerator, and the user will be able
to enter the denominator after the added '/' character.  Fractional
entry works in all modes, but will only display as a fraction after
entry if the fraction display mode is enabled.

Additionally, a symbol display mode was added to show values like pi,
sqrt(2), and sqrt(10).  This mode is vulnerable to rounding errors,
and so may not work reliably.