added docker development environment (#1)
This work adds a developer environment which was setup via docker for
compiling, debugging, and running across different systems.  By default, the run
and debug make targets will have the internal x display pointed to the
host's :0 display.  The dev target can be used to launch the docker
container in interactive mode.